Extracted models to their own files
