+++ src/models/Product.ts
@@ -1,0 +1,42 @@
+import { DataTypes, Model } from 'sequelize';
+import { sequelize } from '../pg';
+
+export class Product extends Model { }
+
+Product.init({
+    id: {
+        type: DataTypes.UUID,
+        primaryKey: true,
+        defaultValue: DataTypes.UUIDV4
+    },
+    name: {
+        type: DataTypes.STRING,
+        allowNull: false
+    },
+    description: {
+        type: DataTypes.TEXT,
+        allowNull: false
+    },
+    price: {
+        type: DataTypes.DECIMAL(10, 2),
+        allowNull: false
+    },
+    img_url: {
+        type: DataTypes.STRING,
+        allowNull: false
+    },
+    categories: {
+        type: DataTypes.ARRAY(DataTypes.STRING),
+        allowNull: false
+    },
+    sizes: {
+        type: DataTypes.ARRAY(DataTypes.STRING),
+        allowNull: false
+    },
+    colours: {
+        type: DataTypes.ARRAY(DataTypes.STRING),
+        allowNull: false
+    }
+}, {
+    sequelize
+});
